FEAT: simplied old file reader for easier refactoring
