--- src/file_readers/spectrum_readers.rs
@@ -1,2 +1,0 @@
-pub mod dda_reader;
-pub mod mini_tdf_reader;
